fix: 좌표 값 타입 수정

+++ src/main/java/com/jigume/dto/goods/GoodsDto.java
@@ -22,9 +22,9 @@ public class GoodsDto {
 
     private Integer deliveryFee;
 
-    private Integer mapX;
+    private Long mapX;
 
-    private Integer mapY;
+    private Long mapY;
 
     private Integer goodsLimitCount;
